bookstore_MySQL/requirements.txt
simplejson~=3.19.2
lxml~=4.9.3
codecov
coverage
flask~=2.0.0
pre - commit
pytest~=7.4.3
PyJWT~=2.8.0
requests~=2.31.0

pymongo~=4.5.0
jieba~=0.42.1
setuptools~=68.2.2